Se agrega metodo para buscar Producto por ID

+++ src/main/java/com/project/gestionPedidos/domain/Producto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Producto {
-    private Integer ProductoId;
+    private Long ProductoId;
     private String NombreProducto;
     private Double PrecioProducto;
     private String DescripcionProducto;
