+++ src/main/java/com/project/gestionPedidos/services/ProductoService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface ProductoService {
     public List<Producto> getProductos();
     public void saveProducto(Producto producto);
-    public Producto getProductoById(Integer ProductoId);
+    public Producto getProductoById(Long ProductoId);
 }
